mtrlz/sql: bump sqlparser dependency
This brings support for 'CREATE DATA SOURCE ... USING SCHEMA REGISTRY
'...'`. The new syntax isn't hooked up to anything yet, though.

For good measure, the new version of the sqlparser library includes a
bunch of upstream changes, which caused some churn in our parsing code.
Notably the upstream changes add support for window functions, which
we'll want to be able to plan eventually.