mtrlz/dataflow: eagerly reclaim insert_mux channel keys
When dropping and recreating a table in quick succession, like

    CREATE TABLE t ...;
    DROP TABLE t;
    CREATE TABLE t ...;

the newly-created table will register a channel with the input mux keyed
on "t". Since input mux channels were previously lazily garbage
collected, there was no guarantee that the input mux channel "t" from
the prior table had been gc'd.

Introduce a close method on the mux that eagerly reclaims names, and
call it when a dataflow is dropped.