mtrlz/dataflow: rudimentary support for CREATE DATA SINK
Data sinks send the stream of updates from a dataflow to an external
system... or they will, at least. At the moment data sinks just spew
output to the console. I'm upstreaming this patch even though it's
half-baked because it causes some rather invasive renames that I'd
rather not rebase.