mtrlz: upgrade sqlparser
Upstream has landed nearly all of our SQL parser patches, but with some
occasional modifications. Upgrade to a sqlparser version that
incorporates these modifications.

Notably this upgrade drops the "DATA" in "DATA SOURCE" and "DATA SINK",
as, upon reflection, "SOURCE" and "SINK" are perfectly clear on their
own, and fit better with "TABLE" and "VIEW".