ci/slt: fix sqllogictest result uploading
The sed script used in this process had gotten a bit out of hand.
Specifically, "\10" is a backreference to the first match, followed by a
literal zero, not a backreference to the tenth match as intended. Teach
sqllogictest to output a JSON summary of results, and use that to
construct the appropriate SQL query.