testdrive: disable garbage collection on Kafka topics
Kafka has a default 7-day retention policy, and testdrive uses
timestamps that appear to be from the 1970s. This results in Kafka
occasionally destroying data while our tests are running, if we are
unlucky enough to be running tests while the Kafka garbage collector is
active. (By default, the Kafka GC runs every 5m.)

This hopefully fixes the CI hangs we've been seeing.