sql: support configuration parameters
PostgreSQL has a notion of configuration parameters, like `DateStyle`
and `client_encoding`. Clients besides `psql` get very upset when these
parameters are not present. This commit introduces support for some of
the fundamental configuration parameters, which brings us close to
supporting pgcli and suppresses a number of warnings when using
`cockroach sql` to connect. (We may want to use one of these alternate
CLIs for better TAIL support in the demo.)

Configuration parameters can be observed with the `SHOW` statement and
changed with the `SET` statement.