mtrlz/glue: inline CommandMeta into DataflowCommand
In order to split out dataflow and sql crates, the glue module needs to
be removed. This is the last step towards that goal. The CommandMeta struct
can easily be inlined into DataflowCommand, as it only contains one field,
and is presently only used by PEEKs.

This commit also removes the glue module, since it no longer contains any
objects.