Handle Semantic Dates
This adds support for Debezium and Kafka's date output format via Chrono's
support for Dates.

Instead of using chrono we could imagine possibly just parsing to an Int32 and
expecting downstream crates to add some semantic layers on top of the Int in
their interpretation.

One downside to the current implementation (afaict) of avro-rs is that there is
no way to expose the various extra complex fields to downstream crates, so
while we could parse "logicalType": "date" to a Date variant, correctly parsing
the various extended date formats that kafka/debezium provide is impossible
without baking them into this code.