Remove the distinction between ScalarType::Interval{Duration,Months}
We aren't allowed to make distinctions in representation (at any level) based
on it, even though the rules are kinda crazy.

Eventually we might want to go closer to postgres' internal representation of
`Months/Days/Hours` to be able to represent "logical" days, rather than numbers
of seconds, and/or combine months and days, but the semantics of that are wild
so I'm leaving it off.