dataflow: move coordinator to its own thread
This is the third commit in what is becoming a long series of commits to
split the coordinator off onto its own thread.

While this commit technically accomplishes the stated goal, the task is
far from complete. The code changes here are the minimum possible, and
one massive defect is that the coordinator spins at 100% CPU. This will
be addressed in a follow-up commit.