dataflow: prepare to move coordinator into new package
This is the fourth commit in a long series of commits to split the
coordinator off onto its own thread.

In the next commit, the coordinator will be moved to the materialized
crate. In preparation for this move, this commit moves several types
that were declared in the dataflow::coordinator module to the
dataflow::server module, since these types will continue to belong in
the dataflow crate even when the coordinator moves out.