dataflow: move coordinator into materialize::queue
This is the fifth commit in a long series of commits to split the
coordinator off onto its own thread.

As promised in the last commit, this commit moves the coordinator out
of the dataflow crate and into the queue module of the materialize
crate. The coordinator and queue will be merged into one entity in the
next commit; this commit is almost entirely code movement to make
things easier to review.