comm: run local switchboards over Unix sockets
Our SQL logic tests generate so many PEEKs that they exhaust the
available outgoing ports, once PEEK traffic is switched to use comm
channels (#487). We will eventually need to use TCP connection pooling
to avoid this problem, since it could happen in production too, but for
now we can just use Unix sockets when traffic is entirely local, like it
is for SQL logic tests.

This is a useful feature in its own right--it's really dumb to open a
TCP connection to localhost, because TCP has a lot of overhead.