sql: allow adding interval and date
The previous code would panic if you tried to add an interval and a
date, rather than a date and an interval. The dataflow layer was
assuming that the interval would be the second argument, while the SQL
layer would happily provide it as the first.

Also, subtracting a date from an interval is ill-defined and should be
outright prohibited.