Add bare minimum prometheus metrics for materialized
This just returns response counts, but it adds the infrastructure to enable more metrics
from inside materialized, and also for us to add some more external prometheus metrics,
coming in a follow-up commit. The only useful query that can be performed because of this
commit `rate(mz_responses_sent_total[10s])` or sub-queries on that (e.g.
`rate(mz_responses_sent_total{kind="select",status="success"})`). The
state-machine-future code makes it difficult to thread the query that we're currently
tracking through to the error case, so currently status="error" always has an empty
"kind".

In addition to the core infra for prometheus inside materialized, this adds an example
local prometheus config in ex/prometheus (it will work if you just `cargo run --bin
materialized`) and also prometheus to the chbench docker-compose setup.

Since I'm not exactly sure what the perf impact of metrics will be I added a flag to turn
them off. It seems like the rust-prometheus library does not store timestamps for
individual `inc()` calls, so I don't think that there is much risk of memory exhaustion
if we are running materialized without prometheus also running, but we could definitely
add more granular metrics in the future that could include arbitrary memory overhead.

Future commits will add prometheus connectors to the timely logging infrastructure.

Part of #317