rust-toolchain: upgrade to beta (temporarily)
Upgrade to the 1.39 beta, temporarily, for async/await support. This
will allow us to incrementally incorporate 1.39 support in the weeks
leading up to the stable release, and report bugs while there's still
time for them to be fixed.

This commit also updates the version of rdkafka, as the previous version
had a bug in its version of bindgen that caused the beta compiler to
overflow its stack.