expr,sql: tighten branch key during decorrelation
When decorrelating a subquery like

    SELECT * FROM t1 WHERE EXISTS (SELECT * FROM t2)

we generate a query plan that involves joining t1 against t2, to
effectively run the inner query once for every row of the outer query.
This is extra computation than is actually required, as in this case the
subquery is not correlated; in other words, the predicate

   EXISTS (SELECT * FROM t2)

does not vary based on the observed row of t1, so the subquery can
simply be executed once, and that result can be applied to every row of
t1.

This generalizes to correlated subqueries like

    SELECT a, b FROM t1 WHERE EXISTS (SELECT * FROM t2 WHERE t1.b = t2.b)

where the inner query only needs to be executed for distinct values of
t1.b, and not distinct values of (t1.a, t1.b), since the value of the
inner query is not dependent on t1.a, just t1.b.

This commit teaches the decorrelator to walk the inner queries,
determine what columns they actually use from the outer query, and use
that to limit the amount of computation that must be performed, as
described above.

Informs #704.