sql: eagerly incorporate predicates when decorrelating
The branch key tightening introduced in #719 had the defect that it was
interfering with the assertive predicate pushdown introduced in #715.
As a result, correlated subqueries could wind up planning large cross
joins, even though a join key was technically available.

This commit not only restores the assertive predicate pushdown behavior,
it enhances to never before seen power: even the five-way cross join in
q21 is finally optimized.

The trick is to take a query that looks like

    SELECT * FROM a, b, c
    WHERE a.a = b.a AND b.c = c.c AND EXISTS (<subquery>)

and rewrite it to look like, essentially:

    SELECT * FROM (
        SELECT * FROM a, b, c WHERE a.a = b.a AND b.c = c.c
    ) WHERE EXISTS (<subquery>)

In the old example, <subquery> is evaluated in the context of `FROM
a, b, c` (i.e., the three-way cross join), and we needed serious
optimization smarts to push down the join key (`a.a = b.a AND ...`) to
both sides of the join.

In the rewritten example, <subquery> is evaluated in the context of
`FROM a, b, c WHERE a.a = b.a AND b.c = c.c` (i.e., the predicates are
applied first), and the branch naturally applies to the post-selection
relation without any kind of complicated optimizations required.

Things get tricky when multiple subqueries are in play, and this patch
doesn't attempt to make optimal decisions. A query like

    SELECT * FROM a, b, c
    WHERE a.a = b.a AND b.c = c.c
    AND EXISTS (<subquery>)
    AND EXISTS (<subquery 2>)

may very well be more efficient if <subquery 2> is evaluated before
<subquery>, but at the moment we simply plan in the order that
subqueries appear.

Fix #750.