pgwire: harmonize message decoding
Ensure each type of message has a dedicated function, even when the
function is trivial, to make the dispatcher in the match easier to
follow. Also name the functions "decode_MESSAGETYPE", rather than
"parse_MESSAGETYPE", which better harmonizes with the name of the
"Decoder" object and avoids the awkwardness of "parse_parse".