comm: support closing MPSC receivers when number of producers is known
MPSC channels in comm have the defect that they do not close when all
possible producers have closed. This is for reasons of efficiency; since
producers can be sent to different processes, refcounting them (e.g.,
with drop implementations) would be unacceptably expensive, as the
refcount would need to be distributed and thus would involve network
accesses.

Instead, introduce a new API where the client specifies the maximum
number of producers that are expected to connect. Once that number of
producers has connected, and once those producers have finished
producing messages, the receiver is closed.