pgwire,sql: support all queries in extended flow
Rework the prepared statement implementation so that any query can be a
"prepared" statement. This is necessary because the pgwire extended
query flow essentially prepares every statement, and many drivers do
this transparently.

Now, a Parse command in the extended flow will cause all the immutable
bits of planning. For example, Parse on a `CREATE VIEW ...` statement
will do everything short of actually installing the view, so that errors
can be surfaced immediately.

When the prepared statement is later the target of an Execute command,
we'll replan the query, this time running both the immutable and mutable
bits. This is true also for `SELECT` statements, which is a bit
unfortunate, as previously the plans were cached. I think this tradeoff
is worth it for now, as the big benefit will come when we can actually
preoptimize plans; the SQL planner itself is quite fast.

As a quick integration test for now, switch testdrive over to using the
extended query flow. This shook out several bugs in our implementation
of the backend extended query flow.

Also, as a bonus, add support for binary encoding of decimals, and also
fix a bug in our binary encoding of timestamps, which was using the
wrong epoch.

Fix #599.
Fix #776.