pgwire: fix simple query protocol
I accidentally broke the simple query protocol in #784; the protocol
state machine was accidentally falling into the extended query flow even
when the simple flow was active due to a simple typo.

Fix that. It's clear that we need some tests for the simple flow, but
I'm going to defer that for the moment, since the incoming JDBC work
will provide hundreds and hundreds of tests for the protocol.