gitignore: don't include editor-specific entries
Don't include editor-specific entries. We don't want to be in the
business of ignoring every user's editor's mess. That's what the global
gitignore file is for.

Also move some doc/user specific entries to doc/user/.gitignore, to
limit the entries in the global gitignore to those that apply globally.

Finally, standardize on glob syntax. Entries that start with a slash
apply to entries at the level of the gitignore; other entries match at
any level beneath the level of the gitignore.