Reduce dataflows created for peeks.
This PR reduces the number of dataflows created for peeks by handling both projections and filters in the peek processsing pipeline, so that if the query is an optional projection around an optional filter around a get, we can process the query without dataflow construction. This should ease the load on the system, and reduce the amount of random nonsense the system has to do in response to relatively easy peek and select queries.