ore,mzd,slt: improve ergonomics of server shutdown
Shutting down a materialized server must be done carefully. First the
command queue transmitter (cmdq_tx) must be shutdown, to send the
shutdown signal. Then we have to wait for the dataflow workers to
shutdown, in order to make sure messages are not still flowing between
the workers and the coordinator. Then we have to wait for the
coordinator to shut down, to make sure nothing is still using the tokio
runtime. And lastly, the tokio runtime can be shutdown.

This was previously handled with a custom drop implementation for
materialized::Server, and a drop-like shutdown function for
sqllogictest::runner::State. These approaches were both flawed;
materialized::Server was forced to use a bunch of unnecessary options so
that it could move out of some fields in its drop implementation, while
sqllogictest was forgetting to call the shutdown method in some modes.

Solve both issues by providing a JoinOnDropHandle that waits for a
thread to shutdown when the handle is dropped. (std::thread::JoinHandle
*detaches* from the thread when dropped.) Then we don't need to provide
a custom drop implementation, but can rely on the field drop order to
handle everything for us.

This has the nice side effect of solving the hang when passing
--rewrite-results to SLT.