sql: split out catalog crate
The catalog will soon be responsible for metadata persistence. Split it
out into its own crate to better encapsulate this responsiblity. It
doesn't have much to do with SQL, after all; the catalog is sufficiently
general that it could be, one day, useful for views defined in datalog,
or in Materialize's IR.