catalog: persist catalog items
Add single-node persistence for the catalog. Restarting a materialized
server will automatically install the same sources and views that were
present when the server was shutdown.

The catalog is stored on disk using SQLite. The default location is
"mzdata/catalog", relative to the CWD of materialized, but can be
configured with the --data-directory command line flag.