dataflow,expr: introduce proper id type
This commit works towards a longstanding goal of a proper ID type for
"dataflow components", where a "dataflow component" is roughly anything
that can be the target of a RelationExpr::Get expression.

Previously, we were using simple string names (or, more recently,
QualNames) across the entire system. But names are a SQL concept, doubly
so when qualified, and the dataflow layer really wants to deal with
something simpler and cheaper. Enter expr::Id.

An expr::Id is either in the local or global namespace; the global
namespace is further subdivided into system and user namespaces. Within
each namespace, identifiers are simply a usize. This makes it dead
simple to allocate IDs, as each layer can use its own ID allocator.
(System global IDs are hardcoded into the binary; user global IDs are
allocated by the catalog; and local IDs are allocated whenever a
RelationExpr is built.)

Fix #993.