repr: remove ScalarType::Time variant
ScalarType::Time is an ancient ScalarType variant (added in 2f2e334cf)
that cannot be constructed, as there is no Datum::Time variant and never
has been. We're not planning on supporting the SQL time type for quite a
while (it has insane semantics around timezones that make it virtually
useless), so just remove ScalarType::Time so that everyone doesn't need
to invent their own way to ignore the variant.