Reduce frontier coordination if stable.
Each worker tracks for each maintained trace the most recent frontier it has reported to the coordinator. Each time around the loop, it means to report any changes back to the coordinator. Right now it determines the changes and reports all of them, independent of which changes are non-empty. This PR changes the behavior to suppress reporting empty change sets, and further to avoid the work in the first place if the frontiers are stable (avoiding some allocation work).