deps: bump to latest master of rust-rdkafka
We no longer need our own fork, as all of our patches have been merged
upstream. Upstream also does not exhibit the bug that was causing
rdkafka to segfault on macOS--a result of our fork using an unreleased
version of the underlying librdkafka C library.

The upgraded version of rust-rdkafka does not depend on bindgen, and so
we no longer have a dependency on clang/libllvm.